Prepare join interview as guest

+++ api/app/channels/interview_channel.rb
@@ -1,6 +1,6 @@
 class InterviewChannel < ApplicationCable::Channel
   def subscribed
-    reject unless interview && Ability.new(current_user).can?(:edit, interview)
+    reject unless interview
     ActiveInterview.add_user(interview, current_user)
     stream_for interview
     send_active_users
@@ -12,22 +12,26 @@ class InterviewChannel < ApplicationCable::Channel
   end
 
   def receive(payload)
-    InterviewChannel.broadcast_to(interview, payload)
-
     case payload['type']
-    when 'start_interview'
-      interview.start!
     when 'interview_ended'
+      return unless can_update_interview?
       interview.end!
+      InterviewChannel.broadcast_to(interview, payload)
       ActiveInterview.remove_all(interview)
     when 'code_updated'
       interview.update_interview!(code: payload['data']['code'])
+      InterviewChannel.broadcast_to(interview, payload)
     when 'draw_updated'
       interview.update_interview!(drawing: payload['data']['elements'])
+      InterviewChannel.broadcast_to(interview, payload)
     when 'title_changed'
+      return unless can_update_interview?
       interview.update_interview!(title: payload['data']['title'])
+      InterviewChannel.broadcast_to(interview, payload)
     when 'question_changed'
+      return unless can_update_interview?
       interview.update_interview!(question_id: payload['data']['question'])
+      InterviewChannel.broadcast_to(interview, payload)
     end
   end
 
@@ -35,6 +39,10 @@ class InterviewChannel < ApplicationCable::Channel
 
   def interview
     @interview ||= Interview.find_by(id: params[:id], status: :started)
+  end
+
+  def can_update_interview?
+    Ability.new(current_user).can?(:update, interview)
   end
 
   def send_active_users
